chore: setup back-end development structure
[finishes #144519111]
