+++ server/index.js
@@ -1,22 +1,15 @@
-import express from 'express';
+/* eslint no-console: "off" */
 import morgan from 'morgan';
-import routes from './routes';
+import app from './app';
 
-const app = express();
 let port = process.env.PORT || 3000;
-
 if (process.env.NODE_ENV === 'development') {
   app.use(morgan('dev'));
   port = 8080;
 }
 
-routes(app);
-
-app.get('/', (req, res) => {
-  res.send({ message: 'Welcome to the future, yeah!' });
+app.listen(port, () => {
+  console.log('Magic happening at: ', port);
 });
 
-app.listen(port);
 
-// Export the server for testing
-export default app;
